Now kernel module is fully functioning in fortran
