+++ fortran/kernel_cdef.f90
@@ -1,28 +1,26 @@
 ! C functions declaration
 interface
 
-    function createKernel_C(a, b) bind(C, name="createKernel")
+    function create_kernel_c(N) bind(C, name="create_kernel")
         use iso_c_binding
         implicit none
-        type(c_ptr) :: createKernel_C
-        integer(c_int), value :: a
-        integer(c_int), value :: b
+        type(c_ptr) :: create_kernel_c
+        integer(c_int), value :: N
     end function
 
-    subroutine deleteKernel_C(K) bind(C, name="deleteKernel")
+    subroutine delete_kernel_c(K) bind(C, name="delete_kernel")
         use iso_c_binding
         implicit none
         type(c_ptr), value :: K
     end subroutine
 
-    subroutine getMatrix_C(K, i, j, n_rows, n_cols) bind(C, name="getMatrix")
+    function get_matrix_entry_c(K, i, j) bind(C, name="get_matrix_entry")
         use iso_c_binding
         implicit none
-        type(c_ptr), value :: K
+        real(c_double) :: get_matrix_entry_c
+        type(c_ptr), intent(in), value :: K
         integer(c_int), value :: i
         integer(c_int), value :: j
-        integer(c_int), value :: n_rows
-        integer(c_int), value :: n_cols
-    end subroutine
+    end function
 
 end interface
